+++ src/main.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom/client';
-import { RouterProvider, createHashRouter } from 'react-router-dom';
+import { RouterProvider, createBrowserRouter } from 'react-router-dom';
 import {
   ApolloClient,
   InMemoryCache,
@@ -9,19 +9,14 @@ import {
 } from '@apollo/client';
 import { setContext } from '@apollo/client/link/context';
 import routes from './routes';
-// import { baphTheme } from './styling/baphTheme';
-// import BaphThemeProvider from './styling/BaphThemeProvider';
 
 const baseURL =
   import.meta.env.VITE_NODE_ENV === 'development'
     ? import.meta.env.VITE_DEV_URI
     : import.meta.env.VITE_PROD_URI;
 
-const router = createHashRouter(routes);
-
-const httpLink = createHttpLink({
-  uri: `${baseURL}/graphql/`
-});
+const router = createBrowserRouter(routes);
+const httpLink = createHttpLink({ uri: `${baseURL}/graphql/` });
 const authLink = setContext((_, { headers }) => {
   const token = localStorage.getItem('baphomet-token');
   return {
@@ -37,13 +32,10 @@ const client = new ApolloClient({
   cache: new InMemoryCache()
 });
 
-// Not using BaphThemeProvider yet, but he can hang out.
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
     <ApolloProvider client={client}>
-      {/* <BaphThemeProvider theme={baphTheme}> */}
       <RouterProvider router={router} />
-      {/* </BaphThemeProvider> */}
     </ApolloProvider>
   </React.StrictMode>
 );